project waiting grade

+++ 01_ft_printf/ft_printhex.c
@@ -6,7 +6,7 @@
 /*   By: lsaba-qu <leonel.sabaquezada@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/25 19:08:05 by lsaba-qu          #+#    #+#             */
-/*   Updated: 2022/11/30 17:57:48 by lsaba-qu         ###   ########.fr       */
+/*   Updated: 2022/11/30 19:24:33 by lsaba-qu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,12 +27,12 @@ int	ft_lenhex(unsigned int num)
 
 int	ft_printhex(unsigned long long num, char *format, int base)
 {
-	if (num >= base)
+	if (num >= (unsigned long long)base)
 	{
 		ft_printhex(num / base, format, base);
 		ft_printhex(num % base, format, base);
 	}
-	else if (num < base)
+	else if (num < (unsigned long long)base)
 		ft_putchar_fd(format[num], 1);
 	return (ft_lenhex(num));
 }
